Move all interface declarations to new `types.ts`
