+++ src/index.ts
@@ -4,4 +4,4 @@ import { cli, command, error } from './waterfall-cli';
 // Exports
 export default cli;
 export { command, error };
-export { Settings } from './utils';
+export { Settings } from './types';
