+++ src/utils.ts
@@ -4,75 +4,13 @@ import fs from 'fs';
 import path from 'path';
 import Fuse from 'fuse.js';
 import ErrorWithoutStack from './error-without-stack';
-
-// Interface declarations
-interface InputObject {
-	command: string;
-	data?: string | number;
-	[index: string]: boolean | string | number | undefined;
-}
-
-// Interface declarations
-export interface AppSettings {
-	name?: string;
-	packageName?: string;
-	version?: string;
-}
-
-export interface CommandSpec {
-	data?: {
-		accepts?: string[];
-		description: string;
-		ignoreFlagsAndOptions?: boolean;
-		required?: boolean;
-		type?: 'integer' | 'float';
-	};
-	description?: string;
-	flags?: {
-		[index: string]: {
-			cascades?: boolean;
-			description?: string;
-			shorthand?: string;
-		};
-	};
-	options?: {
-		[index: string]: {
-			accepts?: string[];
-			cascades?: boolean;
-			description?: string;
-			required?: boolean;
-			shorthand?: string;
-			type?: 'integer' | 'float';
-		};
-	};
-	executeOnCascade?: boolean;
-}
-
-export interface Settings {
-	app: AppSettings;
-	arguments: string[];
-	mainFilename: string;
-	newVersionWarning: {
-		enabled: boolean;
-		installedGlobally: boolean;
-	};
-	onStart?: string | Function;
-	packageFilePath: string;
-	spacing: {
-		after: number;
-		before: number;
-	};
-	usageCommand?: string;
-	verbose?: boolean;
-}
-
-export interface OrganizedArguments {
-	command: string;
-	data?: string | number;
-	flags: string[];
-	options: (string | number)[];
-	values: (string | number)[];
-}
+import {
+	AppSettings,
+	CommandSpec,
+	InputObject,
+	OrganizedArguments,
+	Settings,
+} from './types';
 
 // Helpful utility functions
 export default function utils(currentSettings: Settings) {
